requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backoff==2.1.2
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.7
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
cytoolz==0.12.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
gensim==4.2.0
hdbscan==0.8.28
idna==3.3
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
llvmlite==0.39.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
nltk==3.7
node2vec==0.4.6
notebook==6.4.11
numba==0.56.0
numexpr==2.8.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pause==0.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
PyGithub==1.55
Pygments==2.12.0
pygraphviz==1.9
PyJWT==2.4.0
PyNaCl==1.5.0
pynndescent==0.5.7
pyparsing==3.0.9
pyphen==0.13.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.6.2
requests==2.27.1
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
squarify==0.4.3
srsly==2.4.3
stack-data==0.2.0
tables==3.7.0
tabulate==0.8.9
terminado==0.15.0
textacy==0.12.0
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typer==0.4.1
typing_extensions==4.2.0
umap-learn==0.5.3
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wordcloud==1.8.2.2
wrapt==1.14.1
